Fix routine runner function
Description
-----------

In the generated `Routine` module we use `Application.get_env/2` to retrieve the default routine. It comes as a string but we're appending the function `run()` to it which doesn't work because we should turn the string into a module first.

e.g.

```
== Compilation error in file lib/bot_name/routine.ex ==
** (CompileError) lib/bot_name/routine.ex:23: invalid call "BotName.Routine.HelloWorld".run()
    lib/bot_name/routine.ex:22: (module)
    (stdlib) erl_eval.erl:677: :erl_eval.do_apply/6
    (elixir) lib/kernel/parallel_compiler.ex:208: anonymous fn/4 in Kernel.ParallelCompiler.spawn_workers/6
```

That string comes from configs:

```
config :project_name,
  default_routine: "BotName.Routine.HelloWorld"
```

We can also pass a module in the `Routine` module but the changes here should work for both cases.

Also I have changed the env variable `default_routine` to `<bot-name>_default_routine` once we're aiming to handle multiple bots. In that way we can have a default routine for each bot instead of a single one.